+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-### NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
+### NEXT_VERSION_TYPE=MINOR
 ### NEXT_VERSION_DESCRIPTION_BEGIN
+* Added possibility to specify names of options in a custom variable.
 ### NEXT_VERSION_DESCRIPTION_END
 ## [2.0.1]() (12-09-2019)